demo/requirements.txt
litellm
playwright
python-dotenv
rich